refactor: rename fhe -> fhe_traits
